+++ tket-py/tket/__init__.py
@@ -25,4 +25,4 @@ __all__ = [
 
 # This is updated by our release-please workflow, triggered by this
 # annotation: x-release-please-version
-__version__ = "0.12.7"
+__version__ = "0.12.12"