+++ tket-qsystem/src/lib.rs
@@ -9,7 +9,6 @@ pub mod llvm;
 mod lower_drops;
 pub mod pytket;
 pub mod replace_bools;
-pub mod replace_borrow_arrays;
 
 use derive_more::{Display, Error, From};
 use hugr::{
@@ -25,9 +24,6 @@ use hugr::{
 };
 use lower_drops::LowerDropsPass;
 use replace_bools::{ReplaceBoolPass, ReplaceBoolPassError};
-use replace_borrow_arrays::{ReplaceBorrowArrayPass, ReplaceBorrowArrayPassError};
-use tket::modifier::modifier_resolver::ModifierResolverErrors;
-use tket::modifier::ModifierResolverPass;
 use tket::TketOp;
 
 use extension::{
@@ -48,12 +44,6 @@ pub struct QSystemPass {
     monomorphize: bool,
     force_order: bool,
     lazify: bool,
-    /// Resolve function modifiers.
-    //
-    // TODO: This should be in a default Hugr pass rather than here.
-    // We should move it and deprecate once that's defined.
-    modifier: bool,
-    lower_borrow_arrays: bool,
 }
 
 impl Default for QSystemPass {
@@ -63,8 +53,6 @@ impl Default for QSystemPass {
             monomorphize: true,
             force_order: true,
             lazify: true,
-            modifier: true,
-            lower_borrow_arrays: true,
         }
     }
 }
@@ -75,8 +63,6 @@ impl Default for QSystemPass {
 pub enum QSystemPassError<N = Node> {
     /// An error from the component [ReplaceBoolPass].
     ReplaceBoolError(ReplaceBoolPassError<N>),
-    /// An error from the component [ReplaceBorrowArrayPass].
-    ReplaceBorrowArrayError(ReplaceBorrowArrayPassError<N>),
     /// An error from the component [force_order()] pass.
     ForceOrderError(HugrError),
     /// An error from the component [LowerTketToQSystemPass] pass.
@@ -99,8 +85,6 @@ pub enum QSystemPassError<N = Node> {
     /// [Module]: hugr::ops::Module
     #[display("No function named 'main' in module.")]
     NoMain,
-    /// An error from the component [ModifierResolverPass].
-    ModifierResolverError(ModifierResolverErrors<N>),
 }
 
 impl QSystemPass {
@@ -122,11 +106,6 @@ impl QSystemPass {
             hugr.entrypoint()
         };
 
-        // experimental modifier support
-        if self.modifier {
-            self.resolve_modifier().run(hugr)?
-        }
-
         // passes that run on whole module
         hugr.set_entrypoint(hugr.module_root());
         if self.monomorphize {
@@ -137,19 +116,11 @@ impl QSystemPass {
         }
 
         self.lower_tk2().run(hugr)?;
-        // Only has an effect if there are borrow arrays - should be run before
-        // lazification and drop lowering as no copy discard handler currently exists
-        // for BorrowArrays.
-        if self.lower_borrow_arrays {
-            self.replace_borrow_arrays().run(hugr)?;
-        }
         if self.lazify {
             self.replace_bools().run(hugr)?;
         }
         // We expect any Hugr will have *either* drop ops, or ValueArrays (without drops),
         // so only one of these passes will do anything; the order is thus immaterial.
-        // Drop should come after borrow array replacement so that we don't require a
-        // copy/discard handler for borrow arrays + avoid lowering the discard function.
         self.lower_drops().run(hugr)?;
         self.linearize_arrays().run(hugr)?;
 
@@ -218,10 +189,6 @@ impl QSystemPass {
         ReplaceBoolPass
     }
 
-    fn replace_borrow_arrays(&self) -> ReplaceBorrowArrayPass {
-        ReplaceBorrowArrayPass
-    }
-
     fn constant_fold(&self) -> ConstantFoldPass {
         ConstantFoldPass::default()
     }
@@ -236,10 +203,6 @@ impl QSystemPass {
 
     fn linearize_arrays(&self) -> LinearizeArrayPass {
         LinearizeArrayPass::default()
-    }
-
-    fn resolve_modifier(&self) -> ModifierResolverPass {
-        ModifierResolverPass
     }
 
     /// Returns a new `QSystemPass` with constant folding enabled according to
@@ -281,17 +244,6 @@ impl QSystemPass {
     /// from `tket.qsystem`.
     pub fn with_lazify(mut self, lazify: bool) -> Self {
         self.lazify = lazify;
-        self
-    }
-
-    /// Returns a new `QSystemPass` with modifier resolution enabled according to
-    /// `modifier`.
-    ///
-    /// On by default.
-    ///
-    /// See [`ModifierResolverPass`] for more details.
-    pub fn with_modifier_resolution(mut self, modifier_resolution: bool) -> Self {
-        self.modifier = modifier_resolution;
         self
     }
 }