+++ tket/src/passes.rs
@@ -4,13 +4,20 @@ mod commutation;
 
 pub use commutation::{apply_greedy_commutation, PullForwardError};
 
+pub mod borrow_squash;
+pub use borrow_squash::BorrowSquashPass;
+
 pub mod chunks;
 pub use chunks::CircuitChunks;
+
+pub mod guppy;
+pub use guppy::NormalizeGuppy;
 
 pub mod pytket;
 pub use pytket::lower_to_pytket;
 
 pub mod tuple_unpack;
+#[allow(deprecated)]
 pub use tuple_unpack::find_tuple_unpack_rewrites;
 
 pub mod unpack_container;