+++ tket/src/subcircuit.rs
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 //! Subcircuits of circuits.
 //!
 //! Subcircuits are subgraphs of [`hugr::Hugr`] that use a pre-computed
@@ -908,3 +909,12 @@ mod tests {
         );
     }
 }
+=======
+//! Placeholder file until subcircuit is merged, see
+//! [https://github.com/CQCL/tket2/pull/1054](https://github.com/CQCL/tket2/pull/1054)
+
+mod expression;
+pub use expression::CopyableExpr;
+
+mod interval;
+>>>>>>> pr-1269